added functionality for new C server to broadcast timestamps and stop signals to cameras, tested and confirmed it works. Also updated the recording software to expect a little endian timestamp since both the pis and my PC are little endian

+++ stream_server/include/parse_conf.h
@@ -15,7 +15,7 @@ typedef struct cam_conf {
   char name[CAM_NAME_LEN]; // expected name format rpicamXX\0 where XX is a counter from 00-99
 } cam_conf;
 
-int8_t count_cameras(const char* fpath);
-int8_t parse_conf(cam_conf* confs, int8_t count);
+int count_cameras(const char* fpath);
+int parse_conf(cam_conf* confs, int count);
 
 #endif
